IE<9 fix added in
For IE6, 7 & 8, it is necessary to have a fix to make them do something
sensible with HTML5 and CSS3.  html5shiv basically does the job, but it
is also necessary to redesign the CSS so that the bits that IE needs are
not inside media statements.  Don't I *love* Internet Explorer...